バランス調整＋でかい敵の攻撃と爆発も大きくした

+++ Assets/Scripts/Bullet.cs
@@ -5,7 +5,7 @@ using System.Collections;
 public class Bullet : MonoBehaviour 
 {
 	// 弾の移動スピード
-	public int speed = 10;
+	public float speed = 10;
 
 	// ゲームオブジェクト生成から削除するまでの時間
 	public float lifeTime = 5;
